Modify names Stream and Stock
